cleanup: move Internal to its own file
