+++ runtime/extism.h
@@ -53,6 +53,9 @@ typedef struct ExtismCancelHandle ExtismCancelHandle;
  */
 typedef struct ExtismFunction ExtismFunction;
 
+/**
+ * Internal stores data that is available to the caller in PDK functions
+ */
 typedef struct ExtismCurrentPlugin ExtismCurrentPlugin;
 
 typedef uint64_t ExtismSize;
